Tests clean up
